requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1
django-debug-toolbar==3.1.1
django-jsonfield==1.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
pylint==2.5.3
pytz==2020.1
requests==2.24.0
requests-file==1.5.1
six==1.15.0
sqlparse==0.3.1
tldextract==2.2.3
toml==0.10.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.10
wrapt==1.12.1
